requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
tensorflow==2.17.0